+++ packages/url_launcher/url_launcher_ios/example/integration_test/url_launcher_test.dart
@@ -10,7 +10,7 @@ void main() {
   IntegrationTestWidgetsFlutterBinding.ensureInitialized();
 
   testWidgets('canLaunch', (WidgetTester _) async {
-    UrlLauncherPlatform launcher = UrlLauncherPlatform.instance;
+    final UrlLauncherPlatform launcher = UrlLauncherPlatform.instance;
 
     expect(await launcher.canLaunch('randomstring'), false);