+++ script/tool/lib/src/main.dart
@@ -20,6 +20,7 @@ import 'license_check_command.dart';
 import 'lint_android_command.dart';
 import 'lint_podspecs_command.dart';
 import 'list_command.dart';
+import 'make_deps_path_based_command.dart';
 import 'native_test_command.dart';
 import 'publish_check_command.dart';
 import 'publish_plugin_command.dart';
@@ -58,6 +59,7 @@ void main(List<String> args) {
     ..addCommand(LintPodspecsCommand(packagesDir))
     ..addCommand(ListCommand(packagesDir))
     ..addCommand(NativeTestCommand(packagesDir))
+    ..addCommand(MakeDepsPathBasedCommand(packagesDir))
     ..addCommand(PublishCheckCommand(packagesDir))
     ..addCommand(PublishPluginCommand(packagesDir))
     ..addCommand(PubspecCheckCommand(packagesDir))