+++ packages/webview_flutter/webview_flutter/CHANGELOG.md
@@ -1,9 +1,22 @@
-## 2.4.0
+## 2.7.0
 
-* Updates example app Android compileSdkVersion to 31.
-* Integration test fixes.
 * Added `setCookie` to CookieManager.
 * CreationParams now supports setting `initialCookies`.
+
+## 2.6.0
+
+* Adds support for the `loadRequest` method.
+
+## 2.5.0
+
+* Adds an option to set the background color of the webview.
+
+## 2.4.0
+
+* Adds support for the `loadFile` and `loadHtmlString` methods.
+* Updates example app Android compileSdkVersion to 31.
+* Integration test fixes.
+* Updates code for new analysis options.
 
 ## 2.3.1